+++ frontend/src/views/settings/FileLoading.vue
@@ -47,33 +47,6 @@
       <ToggleSwitch class="item" v-model="localuser.fileLoading.clearAll" @change="updateSettings"
         :name="$t('fileLoading.clearAll')"
         :description="$t('fileLoading.clearAllDescription')" />
-
-      <h3 class="section-title">{{ $t("desktopNotifications.title") }}</h3>
-      <p v-if="!notificationsSupported" class="notification-hint">
-        {{ $t("desktopNotifications.unsupported") }}
-      </p>
-      <template v-else>
-        <p class="notification-hint">{{ $t("desktopNotifications.description") }}</p>
-        <p class="notification-permission">
-          {{ $t("desktopNotifications.permissionStatus", { status: permissionStatusLabel }) }}
-        </p>
-        <button
-          v-if="showPermissionButton"
-          type="button"
-          class="button button--flat permission-button"
-          @click="requestPermission"
-        >
-          {{ $t("desktopNotifications.requestPermission") }}
-        </button>
-        <ToggleSwitch
-          class="item"
-          v-model="desktopNotificationsEnabled"
-          :disabled="!notificationsSupported"
-          @change="onDesktopNotificationsChange"
-          :name="$t('desktopNotifications.enabled')"
-          :description="$t('desktopNotifications.enabledDescription')"
-        />
-      </template>
     </div>
     <div class="card-actions">
       <button
@@ -92,13 +65,6 @@ import { notify } from "@/notify";
 import { state, mutations } from "@/store";
 import { usersApi } from "@/api";
 import ToggleSwitch from "@/components/settings/ToggleSwitch.vue";
-import {
-  getNotificationPermission,
-  isDesktopNotificationsEnabled,
-  isNotificationSupported,
-  requestNotificationPermission,
-  setDesktopNotificationsEnabled,
-} from "@/utils/desktopNotifications";
 
 export default {
   name: "fileLoading",
@@ -109,8 +75,6 @@ export default {
   data() {
     return {
       localuser: { fileLoading: {} },
-      desktopNotificationsEnabled: isDesktopNotificationsEnabled(),
-      notificationPermission: getNotificationPermission(),
     };
   },
   computed: {
@@ -119,22 +83,6 @@ export default {
     },
     active() {
       return state.activeSettingsView === "fileLoading-main";
-    },
-    notificationsSupported() {
-      return isNotificationSupported();
-    },
-    permissionStatusLabel() {
-      switch (this.notificationPermission) {
-        case "granted":
-          return this.$t("desktopNotifications.permissionGranted");
-        case "denied":
-          return this.$t("desktopNotifications.permissionDenied");
-        default:
-          return this.$t("desktopNotifications.permissionDefault");
-      }
-    },
-    showPermissionButton() {
-      return this.notificationPermission !== "granted" && this.notificationPermission !== "unsupported";
     },
   },
   mounted() {
@@ -145,8 +93,6 @@ export default {
     if (this.localuser.fileLoading.downloadChunkSizeMb === undefined || this.localuser.fileLoading.downloadChunkSizeMb === null) {
       this.localuser.fileLoading.downloadChunkSizeMb = 0;
     }
-    this.desktopNotificationsEnabled = isDesktopNotificationsEnabled();
-    this.notificationPermission = getNotificationPermission();
   },
   methods: {
     showTooltip(event, text) {
@@ -158,23 +104,6 @@ export default {
     },
     hideTooltip() {
       mutations.hideTooltip();
-    },
-    async requestPermission() {
-      this.notificationPermission = await requestNotificationPermission();
-      if (this.notificationPermission === "granted") {
-        notify.showSuccessToast(this.$t("desktopNotifications.permissionGranted"));
-      } else if (this.notificationPermission === "denied") {
-        notify.showError(this.$t("desktopNotifications.permissionDeniedHelp"));
-      }
-    },
-    async onDesktopNotificationsChange() {
-      if (this.desktopNotificationsEnabled && this.notificationPermission !== "granted") {
-        await this.requestPermission();
-        if (this.notificationPermission !== "granted") {
-          this.desktopNotificationsEnabled = false;
-        }
-      }
-      setDesktopNotificationsEnabled(this.desktopNotificationsEnabled);
     },
     async updateSettings(event) {
       if (event !== undefined) {
@@ -198,26 +127,6 @@ export default {
   text-align: center;
 }
 
-.section-title {
-  width: 100%;
-  margin: 0.5em 0 0;
-  padding: 0 1em;
-  text-align: left;
-}
-
-.notification-hint,
-.notification-permission {
-  width: 100%;
-  margin: 0;
-  padding: 0 1em 0.5em;
-  opacity: 0.85;
-  font-size: 0.95em;
-}
-
-.permission-button {
-  margin: 0 1em 0.5em;
-}
-
 .settings-number-input {
   display: flex;
   align-items: center;
@@ -236,21 +145,6 @@ export default {
   margin-left: 1em;
   min-width: 2ch;
   text-align: center;
-}
-
-.apply-btn {
-  margin-left: 1em;
-  padding: 0.5em 1em;
-  border: none;
-  background-color: var(--primaryColor);
-  color: white;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 0.9em;
-}
-
-.apply-btn:hover {
-  opacity: 0.9;
 }
 
 .item {